Addresses Review Comments
